Write EER fractionation file for the simple case

+++ src/murfey/server/api.py
@@ -487,7 +487,7 @@ async def process_gain(visit_name, gain_reference_params: GainReference):
 @router.post("/visits/{visit_name}/eer_fractionation_file")
 async def write_eer_fractionation_file(
     visit_name: str, fractionation_params: FractionationParameters
-) -> str:
+) -> dict:
     file_path = (
         Path(machine_config.rsync_basepath)
         / (machine_config.rsync_module or "data")
@@ -495,11 +495,13 @@ async def write_eer_fractionation_file(
         / secure_filename(visit_name)
         / secure_filename(fractionation_params.fractionation_file_name)
     )
+    if file_path.is_file():
+        return {"eer_fractionation_file": str(file_path)}
     with open(file_path, "w") as frac_file:
         frac_file.write(
             f"{fractionation_params.num_frames} {fractionation_params.fractionation} {fractionation_params.dose_per_frame}"
         )
-    return str(file_path)
+    return {"eer_fractionation_file": str(file_path)}
 
 
 @router.post("/visits/{visit_name}/clean_state")
